requirements.txt
fair-esm==2.0.0
graphein==1.7.5
torch==2.1.1
torch_geometric==2.4.0
wget==3.2
torchvision==0.16.1
attrdict==2.0.1